Refactor agent into modules and isolate policy manager
